chore: add support for funcorder
